refactor: clean up imports and move logprobToColor function to utils
